Remove BLD_DIR target, tidied DEPS generation
Realised there's no point having a separate BLD_DIR target because we have to create the sub directory on a per-target basis, so may as well create the intermediate directories at the same time.

Also switched to inline sed, to avoid the temporary file handling. Note the inline flag is different on macos and badly documented - you have to specify a backup file extension (I chose none, to skip the backup), and do so before the command.

Also skipped the DEPS creation and include if we're going to clean anyway.