OMG finally, up and running, with compilation, running and debug.
launch.json is required to allow debugging (either from the Makefile Tools, or the Run and Debug perspective). It started as defaults (type: cppdbg), but switched to CodeLLDB based on this: https://stackoverflow.com/a/67693086/3697870

CodeLLDB doesn't support several of the default parameters, so I've commented them out to remember what the default goal was.

settings.json only has Makefile Tools settings. The extension is not well liked and doesn't do a lot (although apparently generates compile commands for intellisense, but I have not explored). But it does have make/debug/run buttons, which are handy. The config file provides two configurations for the make button, which works nicely for our debug/release setup. And it also configures the run button using the "launchConfigurations". Confusingly, the debug button just seems to use the launch.json settings.