Downplay Makefile Tools methods in favour of Run/Debug perspective
Turns out the cwd is not configurable in Makefile Tools - if it's changed the config is no longer found! Plus it seems to be using cppdgb instead of CodeLLDB, so I don't think it's honouring launch.json after all.

Thus bite the bullet and add a tasks.json, referring to it from launch.json. Now the debugger can be launched from the perspective and a build is automatically request.

The Makefile Tools settings in settings.json could probably be retired soon. I did manage to get it to generate a c_cpp_properties.json file, so it's not completely useless. But the file is very bare bones, so not too useful in this project after all.