Replace the early break checks with a single goto
The logic made heavy use of the trick to detect whether the for loop ended early or not. That is, the loop variable is brought out of scope, and then the loop condition is checked after the loop - if it's still satisfied then the loop ended early due to a break. In for loops that are looking for a single failure/success case, this trick saves have a redundant "success" boolean that is set before break.

The problem is, the logic requires three of these checks stacked on top of each other, and the check is inverted in some (ie. success is loop completed vs success is loop ended early). So it gets very unwieldy following the loop terminations.

Since the success check is just used to continue or break from the outer loop, we can bypass the check by continuing or breaking directly, with a goto. So this variation removes all the loop checks and adds a goto instead, which also means all the loop variables can stay within the loop scope.

Like it better, or not?